First draft for manage directory cleanup
